requirements.txt
# Face Recognition Project — Dependencies
# Install: pip install -r requirements.txt

# Deep Learning
tensorflow>=2.12.0
keras>=2.12.0

# Face Detection
mtcnn>=0.1.1

# Computer Vision
opencv-python>=4.7.0
Pillow>=9.0.0

# Numerical
numpy>=1.24.0

# Machine Learning utilities
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
